+++ CHANGELOG.md
@@ -71,8 +71,9 @@
 - XSPF: improved caching in multiple playlist actions.
 - UI: Added a tip at all input popups for sorting, to specify when 'SORT BY' and similar statements must be used.
 - Folders: improved filtering in some cases with nested folders or showing only specific playlists within a folder.
-- Helpers: updated helpers.
 - UI: optimized repainting to use less resources on statistics mode.
+- UI: changed tooltip for multiple selection to only show the available which work on the entire selection.
+- Helpers: updated helpers.
 - Code cleanup.
 ### Removed
 ### Fixed